Refactor editor component and utilities further, organizing for simpler component files.
